+++ CHANGELOG.md
@@ -4,6 +4,18 @@
 
 ## [Unreleased]
 ### Added
+- New Menu for creating controller, module and service
+- New File Icons for NestJS.
+
+### Fixed
+- Removed deprecated API
+- Updated to the latest template changes
+
+### Changed
+- Updated icon for `nest-cli.json`. 
+
+## [0.0.1]
+### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
 - Initial NestJS CLI Action - Right click anywhere in the nest project and generate nest components. 
 - Bug Fixes